Moved two more SCM mock tests to mock_test.go
