+++ scm/mock_test.go
@@ -7,6 +7,7 @@ import (
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
+	"github.com/google/go-cmp/cmp/cmpopts"
 	"github.com/quickfeed/quickfeed/internal/qtest"
 	"github.com/quickfeed/quickfeed/qf"
 	"github.com/quickfeed/quickfeed/scm"
@@ -209,7 +210,7 @@ var mockTeams = []*scm.Team{
 	},
 }
 
-func TestUpdateMockTeamMembers(t *testing.T) {
+func TestMockUpdateTeamMembers(t *testing.T) {
 	s := scm.NewMockSCMClient()
 	ctx := context.Background()
 	course := qtest.MockCourses[0]
@@ -576,6 +577,45 @@ func TestMockGetIssues(t *testing.T) {
 	}
 }
 
+func TestMockGetIssues2(t *testing.T) {
+	s := scm.NewMockSCMClient()
+	s.Repositories = map[uint64]*scm.Repository{
+		1: {
+			ID:    1,
+			OrgID: 1,
+			Owner: qtest.MockOrg,
+			Path:  qf.StudentRepoName("test"),
+		},
+	}
+
+	ctx := context.Background()
+	opt := &scm.RepositoryOptions{
+		Owner: qtest.MockOrg,
+		Path:  qf.StudentRepoName("test"),
+	}
+
+	wantIssueIDs := []int{}
+	for i := 1; i <= 5; i++ {
+		issue, cleanup := createIssue(t, s, opt.Owner, opt.Path)
+		defer cleanup()
+		wantIssueIDs = append(wantIssueIDs, issue.Number)
+	}
+
+	gotIssueIDs := []int{}
+	gotIssues, err := s.GetIssues(ctx, opt)
+	if err != nil {
+		t.Fatal(err)
+	}
+	for _, issue := range gotIssues {
+		gotIssueIDs = append(gotIssueIDs, issue.Number)
+	}
+
+	less := func(a, b int) bool { return a < b }
+	if equal := cmp.Equal(wantIssueIDs, gotIssueIDs, cmpopts.SortSlices(less)); !equal {
+		t.Errorf("GetIssues() mismatch wantIssueIDs: %v, gotIssueIDs: %v", wantIssueIDs, gotIssueIDs)
+	}
+}
+
 func TestMockDeleteIssue(t *testing.T) {
 	s := scm.NewMockSCMClient()
 	ctx := context.Background()
@@ -869,6 +909,43 @@ func TestMockUpdateIssueComment(t *testing.T) {
 				t.Errorf("%s: expected comment body 'Updated', got '%s'", tt.name, comment)
 			}
 		}
+	}
+}
+
+func TestMockUpdateIssueComment2(t *testing.T) {
+	s := scm.NewMockSCMClient()
+	repo := &scm.Repository{
+		ID:    1,
+		OrgID: 1,
+		Owner: qtest.MockOrg,
+		Path:  qf.StudentRepoName("user"),
+	}
+	s.Repositories = map[uint64]*scm.Repository{
+		1: repo,
+	}
+
+	body := "Issue Comment"
+	opt := &scm.IssueCommentOptions{
+		Organization: repo.Owner,
+		Repository:   repo.Path,
+		Body:         body,
+	}
+
+	issue, cleanup := createIssue(t, s, opt.Organization, opt.Repository)
+	defer cleanup()
+
+	opt.Number = issue.Number
+	// The created comment will be deleted when the parent issue is deleted.
+	commentID, err := s.CreateIssueComment(context.Background(), opt)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// NOTE: We do not currently return the updated comment, so we cannot verify its content.
+	opt.Body = "Updated Issue Comment"
+	opt.CommentID = commentID
+	if err := s.UpdateIssueComment(context.Background(), opt); err != nil {
+		t.Fatal(err)
 	}
 }
